Move `USE_TZ` to `aiida.manage.configuration.settings`
This allows to remove `aiida.settings` and means there is now only one
single place to configure instance wide settings.
